requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
dateparser==0.7.4
httplib2==0.18.1
idna==2.10
invoice2data==0.3.5
numpy==1.19.0
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.3.0.36
pdf2image==1.13.1
pdfkit==0.6.1
Pillow==7.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pypiwin32==223
pytesseract==0.3.4
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
PyYAML==5.3.1
regex==2020.5.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tzlocal==2.1
unicode==2.7
Unidecode==1.1.1
urllib3==1.25.9
openpyxl==3.0.4
